Extract annotated requirements builder
